+++ src/react-components/src/components/01-atoms/Button/ButtonTypes.tsx
@@ -3,3 +3,9 @@ export enum ButtonTypes {
   Filled = "filled",
   Link = "link"
 }
+
+export enum ButtonIconPositions {
+  Left = "icon-left",
+  Right = "icon-right",
+  JustifyRight = "justify-right"
+}