+++ src/react-components/CHANGELOG.md
@@ -14,6 +14,8 @@ Currently, this repo is in Prerelease.  When it is released, this project will a
 - Deprecated `SearchResultsList`, `IconLinkList` and `EditionsList` in favor of `UnorderedList`
 - `type` in `Button` now corresponds to html button types.  `buttonType` in button corresponds to the visual categories of buttons (eg: `filled`, `outline`)
 - `subtitle` in `SearchResultsItem` changed to `subtitleContent` to take a JSX
+- `Label` and `Button` only accept children when called directly
+
 
 ### Added: 
 - `Checkbox`